Rename libcurl_ydb_wrapper to libcurl for ydb_env_set
ydb_env_set uses the name of the file to set-up the xc variable. So
previously, since we used `libcurl_ydb_wrapper`, it ydb_env_set set-up
the xc file to use that, resulting in a package name of
`libcurl_ydb_wrapper`. But you cannot use `$&libcurl_ydb_wrapper` in M,
as it's an invalid identifier.

In the source code of this package and in documentation, I have always
used `$&libcurl`; so it's simplest to just rename and let ydb_env_set do
the rest of the work. It now works well.