Fix empty payload issue and fix one TLS test.
While developing for
https://gitlab.com/YottaDB/Util/YDB-Web-Server/-/issues/87, I found that
the libcurl package had a problem with empty payloads. This was because
it didn't send a payload if it was empty. This commit fixes this. Tested
by TPAY0 test.

TCERT2 now fails due to openssl commands not working as expected. Now
fixed and test passes, but I had no luck with TCERT1. Next time.